✨ feat: Update Header
Update Header component to display current route name  This commit updates the Header component to display the name of the current route in the heading, instead of the logo and "Keep" text. It achieves this by using the `useLocation` hook from the `react-router-dom` library to get the current pathname, and then updating the text of the Heading component accordingly.  The `pathName` constant is used to store the capitalized name of the current path, after removing the first "/" character. If there is no path name, the default text "Keep" is displayed.

This feature improves the usability of the web application, as it provides a clearer indication of the user's current location within the application.